Tweaking the scale of the charts to use the full page width. Applied a max width to the images so they do not roll off of the page.

+++ src/main/java/com/github/onsdigital/babbage/highcharts/HighChartsExportClient.java
@@ -49,7 +49,7 @@ public class HighChartsExportClient {
         return getImage(chartConfig, width, null);
     }
 
-    public InputStream getImage(String chartConfig, Integer width, Integer scale) throws IOException {
+    public InputStream getImage(String chartConfig, Integer width, Double scale) throws IOException {
         System.out.println("Calling Highcharts export server");
         List<NameValuePair> postParameters = new ArrayList<>();
         postParameters.add(new BasicNameValuePair("options", chartConfig));
